Add resource-based concurrency control
This is based on the system in saurabhnanda/odd-jobs#55, and aims to
address saurabhnanda/odd-jobs#38, where jobs need to have some limited
access to resources that controls how many can run simultaneously.

Unlike that PR, this implements a system where jobs can require access
to 0 or more resources, with different amounts of usage. This is because
of a business need for jobs to be able to require multiple resources.

The implementation is intended to have no performance impact on existing
code wherever the user has not opted in to resource-based concurrency.
This is done by having parallel implementations wherever necessary that
switch based on the concurrency control chosen.

This subsumes and replaces the job-type-based concurrency control,
because it is possible to model that system using job types as
resources.