Isolate and expose poll run job
this was previousy an internal let binding to
the job poller,
but it's usefull because it allows us to
run a single job.

This will allows us to make determinstic job
running tests.
Since we can just call this function by hand
rather then relying on thread delays

Fix compile errors

Use more fine grained connection pool usage

data.pool has already a built in resource connection caching
mechanism:
https://hackage.haskell.org/package/resource-pool-0.3.1.0/docs/Data-Pool.html#t:PoolConfig

So it's not wise to keep the connection open for the duration
of the worker process (even if we're waiting for a new job).
Furthermore this improves the API of pollRunJob because
we no longer have to pass in a connection,
and use the global connection pool instead.

add comment indicating intent

Return the async as well